requirements.txt
absl-py==0.10.0
opt_einsum
keras_preprocessing
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
imutils==0.5.3
ipykernel==5.3.4
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3
matplotlib==3.3.2
mistune==0.8.4
mtcnn==0.1.0
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.18.5
oauthlib==3.1.0
opencv-contrib-python-headless
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.3
pandocfilters
parso
pickleshare
Pillow
prometheus-client
prompt-toolkit
protobuf
pyasn1
pyasn1-modules
pycparser
Pygments
pyparsing
pyrsistent
python-dateutil
python-engineio
python-socketio
pytz==2020.1
pywinpty
PyYAML
pyzmq
qtconsole
QtPy
requests
requests-oauthlib
rsa
scikit-learn
scipy
Send2Trash
six
sklearn
statsmodels
tabula-py
tensorboard
tensorboard-plugin-wit
tensorflow-cpu
tensorflow-estimator
termcolor
terminado
testpath
threadpoolctl
tornado
pyside2
tqdm
urllib3
waitress
wcwidth
webencodings
Werkzeug
widgetsnbextension
wincertstore
wrap
WTForms
XlsxWriter
zipp
tflearn